muchos cambios

+++ MythicOdessey2/Assets/Scripts/Player/PlayerMagicBar.cs
@@ -7,13 +7,13 @@ public class PlayerMagicBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxMagic(int magic)
+    public void SetMaxMagic(float magic)
     {
         slider.maxValue = magic;
         slider.value = magic;
     }
 
-    public void SetMagic(int magic)
+    public void SetMagic(float magic)
     {
         slider.value = magic;
     }
